getExchange function in a new file.
